requirements.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
absl-py==2.4.0
asgiref==3.11.0
astunparse==1.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
iniconfig==2.3.0
keras==3.13.2
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.2
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==3.0.0
pillow==12.1.0
pluggy==1.6.0
protobuf==6.33.5
Pygments==2.19.2
PyJWT==2.11.0
pytest==9.0.2
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.3.2
setuptools==80.10.2
six==1.17.0
sqlparse==0.5.5
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wheel==0.46.3
wrapt==2.1.1
